GARD Chatbot App/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.21.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.33
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.38.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.76.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.14
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.45.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
